fix(saver): future is also cancelled during timeout
when using timeout, it is actually cancelling current
task, which eventually cancels the future which current
task is awaiting on. after shard once due to timeout,
the future is not re-created. when awaiting the future
again, it raises CancelledError. this time, Timeout
is not capturing it, causing the whole t_sharder Task
cancelled.
to fix it, use `asyncio.shield` to protect the future
from cancelling.